+++ my-app/src/Game/Core/Items/Item.ts
@@ -1,0 +1,8 @@
+import { SpriteBox } from "../StartData/StartData";
+
+export interface Item {
+    type: "string",
+    sprite: "string",
+    value: number,
+    spriteBox: SpriteBox,
+}
