--- my-app/src/Game/Core/Items/Items.ts
@@ -1,1 +1,0 @@
-export const inv = " "
